+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project follows [Semantic Versioning](http://semver.org/).
 ## [2.1.0] - 2020-11-19
 ### Added
 - Reuse selection criteria values from the Patient selection area
+### Fixed
+- Observation list does not match term after selection
 
 ## [2.0.0] - 2020-11-06
 ### Added